# dist/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Filmes",
        "description": "API para gerenciamento de filmes",
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
            "get": {
                "tags": ["Movies"],
                "summary": "Obter todos os filmes",
                "responses": {
                    "200": {
                        "description": "Retorna uma lista de filmes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "tags": ["Movies"],
                "summary": "Atualizar um filme existente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme a ser atualizado",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Filme atualizado com sucesso"
                    },
                    "404": {
                        "description": "Filme não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": ["Movies"],
                "summary": "Remover um filme",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme a ser removido",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filme removido com sucesso"
                    },
                    "404": {
                        "description": "O filme não foi encontrado"
                    },
                    "500": {
                        "description": "Não foi possível remover o filme"
                    }
                }
            }
        },
        "/movies/{genreName}": {
            "get": {
                "tags": ["Movies"],
                "summary": "Filtra os filmes por gênero",
                "parameters": [
                    {
                        "name": "genreName",
                        "in": "path",
                        "required": true,
                        "description": "Gênero do filme",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna uma lista de filmes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Não foi possível remover o filme"
                    }
                }
            }
        },
        "/genres": {
            "get": {
                "tags": ["Genres"],
                "summary": "Obter lista dos gêneros",
                "responses": {
                    "200": {
                        "description": "Retorna uma lista de gêneros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenreList"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": ["Genres"],
                "summary": "Cadastrar um novo gênero",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Genre"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gênero adicionado com sucesso"
                    },
                    "400": {
                        "description": "O nome do gênero é obrigatório"
                    },
                    "404": {
                        "description": "Gênero não encontrado"
                    },
                    "409": {
                        "description": "Já existe um gênero com esse nome"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/genres/{id}": {
            "put": {
                "tags": ["Genres"],
                "summary": "Atualizar um gênero existente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do gênero à ser atualizado",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Genre"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gênero atualizado com sucesso"
                    },
                    "400": {
                        "description": "O nome do gênero é obrigatório"
                    },
                    "404": {
                        "description": "Gênero não encontrado"
                    },
                    "409": {
                        "description": "Já existe um gênero com esse nome."
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": ["Genres"],
                "summary": "Remover um gênero",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do gênero a ser removido",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gênero removido com sucesso"
                    },
                    "404": {
                        "description": "O Gênero não foi encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID do filme"
                    },
                    "title": {
                        "type": "string",
                        "description": "Título do filme"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "Gênero do filme"
                    },
                    "genres": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do gênero"
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do gênero"
                            }
                        }
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Idioma do filme"
                    },
                    "languages": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do Idioma"
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do Idioma"
                            }
                        }
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de Oscars ganhos pelo filme"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": [
                    "title",
                    "genre_id",
                    "language_id",
                    "oscar_count",
                    "release_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título do filme"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "ID do gênero do filme"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "ID do idioma do filme"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de Oscars ganhos pelo filme"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
                }
            },
            "GenreList": {
                "type": "object",
                "required": ["id", "name"],
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID do gênero"
                    },
                    "name": {
                        "type": "string",
                        "description": "Nome do gênero"
                    }
                }
            },
            "Genre": {
                "type": "object",
                "required": ["name"],
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Nome do gênero"
                    }
                }
            }
        }
    }
}